feat(db): migrate from Planetscale Postgres to Cloudflare D1
- schema.ts ported to sqlite-core (autoincrement PKs, text json, epoch
  int timestamps, 0/1 booleans); fresh 0000_d1_baseline migration
- drizzle.config dialect: sqlite
- alchemy.run.ts: bind D1Database (auto read replication, apac primary),
  drop Hyperdrive + Singapore placement
- DB client layer -> drizzle-orm/d1; drop postgres-js + DIRECT_DB_URL
- rewrite interactive transactions to sequential ops (D1 has none):
  checkout, payment confirmation, purchase CRUD, ai-purchase import
- ilike -> like; drop Postgres ::int casts; NOW() -> unixepoch()

+++ packages/api/src/queries/categories.ts
@@ -93,7 +93,7 @@ export const categoryQueries = {
 		},
 
 		async getAllCategories() {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			return db()
 				.select({
@@ -135,7 +135,7 @@ export const categoryQueries = {
 		},
 
 		async getAllCategoriesWithStock() {
-			const productCount = sql<number>`count(${ProductsTable.id})::int`;
+			const productCount = sql<number>`count(${ProductsTable.id})`;
 
 			return db()
 				.select({
